refactor: Add predictor_vru.cpp and utils.cpp to map_based_prediction_node
